Move cloudwatch fake in paas metrics to cw pkg
