+++ tools/metrics/pkg/cloudwatch/fakes/fake_cloudwatch_client.go
